Interest version 5 (#71)
* [WIP] Interest version 5

This is a first draft about how we could move on to the interest version 5 method (difference between total interest). There's plenty of work to do, but it would be interesting to get some early feedback on the design before filling in the rest of the details.

* Continue refactor for interest rate v5

* Continue to clean up interest v5 code

* Continue to clean up new interest formula

* Rework to better fit existing model and allow upgradability

* Calculator -> Model

* Get test cases running again and change migrations, because they are disappointing.

* Sprinkle magic "saveBlockInterest" dust everywhere.

* start to pass test cases, change block scale

* Fix InterestRateStorage tests

* Start to test actual MoneyMarket usage

* begin to fix all test cases

* Minor cleanup, all but one test passing

* Add a new deploy function for v8

* Fix broken test cases and address.js for new contracts

* Minor cleanup and specify uint sizes

+++ contracts/storage/BorrowStorage.sol
@@ -14,10 +14,10 @@ import "../base/Owned.sol";
   */
 contract BorrowStorage is Owned, Allowed, ArrayHelper {
     address[] public borrowableAssets;
-    uint public minimumCollateralRatio;
+    uint256 public minimumCollateralRatio;
 
     event NewBorrowableAsset(address asset);
-    event MinimumCollateralRatioChange(uint newMinimumCollateralRatio);
+    event MinimumCollateralRatioChange(uint256 newMinimumCollateralRatio);
 
     /**
       * @notice `addBorrowableAsset` adds an asset to the list of borrowable assets
@@ -42,7 +42,7 @@ contract BorrowStorage is Owned, Allowed, ArrayHelper {
       * @dev used like this to gate borrow creation: borrower_account_value * minimumCollateralRatio >= borrow_amount
       * @return success or failure
       */
-    function setMinimumCollateralRatio(uint minimumCollateralRatio_) public returns (bool) {
+    function setMinimumCollateralRatio(uint256 minimumCollateralRatio_) public returns (bool) {
         if (!checkOwner()) {
             return false;
         }
